pyproject.toml
[project]
name = "new_pipline"
version = "0.1.0"
description = "Box generation pipeline"
requires-python = ">=3.8"
dependencies = [
    "einops>=0.8.1",
    "gateloop-transformer>=0.2.5",
    "kornia>=0.7.3",
    "matplotlib>=3.7.5",
    "numpy>=1.24.4",
    "pandas>=2.0.3",
    "pillow",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.3.2",
    "scipy>=1.10.1",
    "seaborn>=0.13.2",
    "swanlab>=0.6.7",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "x-transformers>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."] 
